requirements.txt
Flask==1.1.1
gunicorn==20.0.4
boto3==1.10.50
pandas==1.2.4
pandas_datareader
numpy
matplotlib
yfinance
google-cloud-storage==1.14.0
gcsfs==0.2.1
blob

